Refactor the MavenCoordinate class and move its tests to core
