+++ core/src/test/java/eu/fasten/core/data/opal/MavenCoordinateTest.java
@@ -16,18 +16,13 @@
  * limitations under the License.
  */
 
-package eu.fasten.analyzer.javacgopal.data;
+package eu.fasten.core.data.opal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.ArrayList;
 import java.util.Collections;
 
-import eu.fasten.core.data.opal.MavenCoordinate;
-import eu.fasten.core.data.opal.exceptions.MissingArtifactException;
-import eu.fasten.core.maven.utils.MavenUtilities;
 import org.junit.jupiter.api.Test;
 
 class MavenCoordinateTest {
@@ -72,44 +67,5 @@ class MavenCoordinateTest {
     void toURL() {
         var coordinate = MavenCoordinate.fromString("GroupID:ArtifactID:Version", "jar");
         assertEquals("repo/GroupID/ArtifactID/Version/ArtifactID-Version.jar", coordinate.toProductUrl("repo/", "jar"));
-    }
-
-    // ------------------
-    // Maven Resolver Tests
-    // ------------------
-
-    @Test
-    void downloadJarEmptyRepos() {
-        var coordinate = new MavenCoordinate("group", "artifact", "version", "jar");
-        coordinate.setMavenRepos(new ArrayList<>());
-        var resolver = new MavenCoordinate.MavenResolver();
-
-        assertThrows(MissingArtifactException.class, () -> resolver.downloadArtifact(coordinate, null));
-    }
-
-    @Test
-    void downloadJarWrongRepos() {
-        var coordinate = new MavenCoordinate("group", "artifact", "version", "jar");
-        coordinate.setMavenRepos(new ArrayList<>(Collections.singletonList("repo")));
-        var resolver = new MavenCoordinate.MavenResolver();
-
-        assertThrows(MissingArtifactException.class, () -> resolver.downloadArtifact(coordinate, null));
-    }
-
-    @Test
-    void downloadJarPomPackaging() {
-        var coordinate = new MavenCoordinate("group", "artifact", "version", "pom");
-        coordinate.setMavenRepos(new ArrayList<>(Collections.singletonList("repo")));
-        var resolver = new MavenCoordinate.MavenResolver();
-
-        assertThrows(MissingArtifactException.class, () -> resolver.downloadArtifact(coordinate, null));
-    }
-
-    @Test
-    void downloadJarWrongPackaging() throws MissingArtifactException {
-        var coordinate = new MavenCoordinate("org.slf4j", "slf4j-api", "1.7.30", "wrongPackagingType");
-        var resolver = new MavenCoordinate.MavenResolver();
-
-        assertNotNull(resolver.downloadArtifact(coordinate, MavenUtilities.MAVEN_CENTRAL_REPO));
     }
 }
